+++ Storage/FileHolder.cpp
@@ -80,9 +80,8 @@ uint16_t FileHolder::fgetc16be() {
 void FileHolder::ensure_file_is_at_least_length(long length) {
 	fseek(file_, 0, SEEK_END);
 	long bytes_to_write = length - ftell(file_);
-	if(bytes_to_write > 0)
-	{
-		uint8_t *empty = new uint8_t[bytes_to_write];
+	if(bytes_to_write > 0) {
+		uint8_t *empty = new uint8_t[(size_t)bytes_to_write];
 		memset(empty, 0, (size_t)bytes_to_write);
 		fwrite(empty, sizeof(uint8_t), (size_t)bytes_to_write, file_);
 		delete[] empty;
